Added comment rendering to ParameterLines, setting up project with Poetry
