+++ handcalcs/render.py
@@ -1,5 +1,5 @@
-import handcalcs as hand
-import handcalcs.numpy_and_pandas_patches as patches
+from . import handcalcs as hand
+from . import numpy_and_pandas_patches as patches
 from IPython.core.magic import (Magics, magics_class, cell_magic, register_cell_magic)
 from IPython import get_ipython
 from IPython.core.magics.namespace import NamespaceMagics
