+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "handcalcs"
+version = "0.1.0"
+description = "A library and Jupyter extension for automatically rendering your Python calculation code into beautiful latex, with symbolic substitution steps."
+authors = ["Connor Ferster"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+ipython = "^7.15.0"
+ipykernel = "^5.3.0"
+pyparsing = "^2.4.7"
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.4.3"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
